Added missing headers inside request
