+++ Projects/Claims/Sources/Helpers/Network/OdysseyNetworkClient+FileUploader.swift
@@ -1,6 +1,5 @@
 import Flow
 import Foundation
-import hGraphQL
 
 public protocol FileUploaderClient {
     func upload(flowId: String, file: UploadFile) throws -> Future<UploadFileResponseModel?>
@@ -28,50 +27,5 @@ extension OdysseyNetworkClient: FileUploaderClient {
                 }
             return NilDisposer()
         }
-    }
-}
-
-enum OdysseyRequest {
-    case uploadAudioFile(flowId: String, file: UploadFile)
-
-    var baseUrl: URL {
-        return Environment.current.odysseyApiURL
-    }
-
-    var methodType: String {
-        switch self {
-        case .uploadAudioFile:
-            return "POST"
-        }
-    }
-
-    var asRequest: Future<URLRequest> {
-        var request: URLRequest!
-        switch self {
-        case let .uploadAudioFile(flowId, file):
-            var baseUrlString = baseUrl.absoluteString
-            baseUrlString.append("/api/flows/\(flowId)/audio-recording")
-            let url = URL(string: baseUrlString)!
-            let multipartFormDataRequest = MultipartFormDataRequest(url: url)
-            multipartFormDataRequest.addDataField(
-                fieldName: file.name,
-                fileName: file.name,
-                data: file.data,
-                mimeType: file.mimeType
-            )
-            request = multipartFormDataRequest.asURLRequest()
-        }
-        request.httpMethod = self.methodType
-        return Future { completion in
-            TokenRefresher.shared.refreshIfNeeded()
-                .onValue {
-                    if let token = TokenRefresher.shared.getAccessToken() {
-                        request.setValue(token, forHTTPHeaderField: "Authorization")
-                    }
-                    completion(.success(request))
-                }
-            return NilDisposer()
-        }
-
     }
 }
